@@ -19,7 +19,6 @@ setup(
     author="Datadog, Inc.",
     author_email="dev@datadoghq.com",
     classifiers=[
-        "Programming Language :: Python :: 2.7",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
@@ -27,15 +26,12 @@ setup(
     ],
     keywords="datadog aws lambda layer",
     packages=["datadog_lambda"],
-    python_requires=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, !=3.5.*, <4",
+    python_requires=">=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, !=3.5.*, <4",
     install_requires=[
         "datadog==0.41.0",
         "ddtrace==0.53.0",
         "wrapt==1.11.2",
-        # If building for Python 3, use the latest version of setuptools
-        "setuptools>=54.2.0; python_version >= '3.0'",
-        # If building for Python 2, use the latest version that supports Python 2
-        "setuptools>=44.1.1; python_version < '3.0'",
+        "setuptools>=54.2.0; python_version >= '3.0'"
     ],
     extras_require={
         "dev": [
