@@ -1,12 +1,8 @@
 import os
-import sys
 import unittest
 import httpretty
 
-try:
-    from unittest.mock import patch
-except ImportError:
-    from mock import patch
+from unittest.mock import patch
 
 from datadog_lambda.extension import (
     is_extension_running,
@@ -20,9 +16,6 @@ def exceptionCallback(request, uri, headers):
 
 
 class TestLambdaExtension(unittest.TestCase):
-    # do not execute tests for Python v2.x
-    __test__ = sys.version_info >= (3, 0)
-
     @patch("datadog_lambda.extension.EXTENSION_PATH", os.path.abspath(__file__))
     def test_is_extension_running_true(self):
         httpretty.enable()
